src/resources/sagemaker/factorization-machines-classification/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
click==7.0                # via -r requirements.in, flask
flask==1.1.1              # via -r requirements.in
gunicorn==19.9.0          # via -r requirements.in
humanfriendly==4.18       # via -r requirements.in
itsdangerous==1.1.0       # via flask
jinja2==2.10.1            # via flask
markupsafe==1.1.1         # via jinja2
python-box==3.4.2         # via -r requirements.in
pyyaml==5.1.1             # via -r requirements.in
werkzeug==1.0.1           # via -r requirements.in, flask

numpy==1.19.5
Pillow>=8.0
mxnet
